repo_results_2/HumanEval/72/task_72.py
def will_it_fly(q, w):
	''' 
	This function checks if a given list 'q' is palindromic and if the sum of its elements is less than or equal to a given weight 'w'.
	If both conditions are met, the function returns True, indicating that the object will fly. Otherwise, it returns False.
	'''
	# Check if the list is palindromic
	if q != q[::-1]:
		return False
	# Check if the sum of the elements is less than or equal to the maximum weight
	if sum(q) > w:
		return False
	# If both conditions are met, return True
	return True
